Solved tptn

+++ 11001~12000/11050 이항계수1.py
@@ -9,7 +9,7 @@ def factorial(n):
     return ans
 
 def e_hang(n,k):
-    return factorial(n) / factorial(k) / factorial(n-k)
+    return int(factorial(n) / factorial(k) / factorial(n-k))
 
 n , k = map(int,input().split())
